Specify the IKeyboard and IUserKeyboard interfaces
The IUserKeyboard is what the user can use to manipulate the state of the
Chip-8 virtual machine, while the IKeyboard is what the core of the virtual
machine uses to interpret said input. These are two separate interfaces that
don't inherit one from the other to keep concerns fully isolated from each
other.