Add some changes to ITimer interface, specify ITimerObserver one
Implement (again) the observer pattern - in the original Kotlin CHIPEIT prokect
the ITimerObserver is actually called a 'ISwitchObserver', which hides its
intended use. I found it hard to read when coming back to the project after
a few months (same with IClockDivider - IClockDividerObserver, the later was
actually called IClockObserver), so the rename was mandatory.